% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{PHREEQC keywords}
\format{A character vector}
\source{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-5.htm#50528257_46140}
}
\usage{
data(keywords)
}
\description{
Commands that signify the beginning of a PHREEQC code block and its type
}
\examples{
data(keywords)
print(keywords)

}
\references{
David L. Parkhurst and C.A.J. Appelo: Description of Input
and Examples for PHREEQC Version 3--A Computer Program for Speciation,
Batch-Reaction, One-Dimensional Transport,
and Inverse Geochemical Calculations
}
\keyword{datasets}
