% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_detect_keyword}
\alias{phr_detect_keyword}
\title{Detect keywords}
\usage{
phr_detect_keyword(x)
}
\arguments{
\item{x}{A string, usually a line from a .pqi-file}
}
\value{
A logical vector indicating whether or not the input was a keyword
}
\description{
Detect keywords
}
\details{
This function is called by \code{\link{phr_tidy_PHREEQC}}.
It is not meant to be exported or to be used anywhere else, yet.
}
