% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_end}
\alias{phr_end}
\title{Insert END block}
\usage{
phr_end()
}
\description{
This function has no arguments. It creates an END block
that ends the data input for a simulation. After this keyword is read by
PHREEQC, the calculations described by the input for the
simulation are performed and the results printed.
Additional simulations may follow in the input data set,
each in turn will be terminated with an END keyword or the end of the file.
}
\examples{
phr_end()

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/html/final-37.html#pgfId-254414}

A \code{\link[=phr_input_section]{phr_input_section()}}
}
