% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_equilibrium_phases}
\alias{phr_equilibrium_phases}
\title{Specify a list of phases for which equilibria should be established}
\usage{
phr_equilibrium_phases(.number = 1, .name = NA, ...)
}
\arguments{
\item{.number}{The number of the component}

\item{.name}{The name of the component}

\item{...}{key/value pairs in the form Component = c(low, high), where low and high
correspond to saturation indicies.}
}
\value{
A \link{phr_input_section}
}
\description{
Specify a list of phases for which equilibria should be established
}
\examples{
# example 2 from the manual
phr_run(
  phr_solution(), # pure water
  phr_equilibrium_phases(Gypsum = c(0, 1), Anhydrite = c(0, 1)),
  phr_reaction_temperature(low = 25, high = 50, steps = 10),
  phr_selected_output(temperature = TRUE, si = c("Gypsum", "Anhydrite"))
)

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-13.htm#50528257_61207}
}
