% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_mix}
\alias{phr_mix}
\title{Specify a MIX of solutions}
\usage{
phr_mix(.number = 1, .name = "", ...)
}
\arguments{
\item{.number}{Number of the component}

\item{.name}{Name of the component}

\item{...}{Further arguments}
}
\value{
A \link{phr_input_section}
}
\description{
Specify a MIX of solutions
}
\examples{
solution_1 <- phr_solution(.number = 1, Na = 1, Cl = 1)
solution_2 <- phr_solution(.number = 2, K = 1, Cl = 1)
phr_mix("1" = 0.25, "2" = 0.75) # Referring to SOLUTION ".number" in quotes

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-27.htm#50528257_23725}
}
