% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_parse_PHREEQC}
\alias{phr_parse_PHREEQC}
\title{Translate PHREEQC to tidyphreeqc}
\usage{
phr_parse_PHREEQC(x)
}
\arguments{
\item{x}{A list of strings that constitute a PHREEQC input block}
}
\value{
A \code{\link{phr_input_section}}
}
\description{
Take PHREEQC-code and fill it into \code{\link{phr_input_section}}'s
}
\details{
This function is called by \code{\link{phr_tidy_PHREEQC}}.
It is not meant to be exported or to be used anywhere else, yet.
}
