% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_reaction}
\alias{phr_reaction}
\title{Specify a REACTION}
\usage{
phr_reaction(.number = 1, .name = "", Reaction_amount,
  Linear_steps = NA, units = "moles", ...)
}
\arguments{
\item{.number}{Number of the component}

\item{.name}{Name of the component}

\item{Reaction_amount}{Amount of components added per step}

\item{Linear_steps}{Number of times a \code{Reaction_amount} shall be added}

\item{units}{Character vector of length = 1. Either "moles", "millimoles" or "micromoles"}

\item{...}{Further arguments}
}
\value{
A \link{phr_input_section}
}
\description{
Define irreversible reactions that transfer a specified amount of elements
into your SOLUTION during a batch reaction simulation.
}
\details{
The steps at which components are added to the solution can be
defined either explicitly or implicitly. For an explicit definition provide
\code{Reaction_amoun} with a vector where each element refers to one
 amount (in "units") at the step of the elements index. For the implicit
 definition, \code{Reaction_amount} needs to be given as a single numeric and
 \code{Linear_steps} is set to the number of times that amount is added to
 the SOLUTION to cause a REACTION.
}
\examples{
#Explicit definition:
phr_reaction(Anhydrite = 1, Reaction_amount = c(1, 1, 2, 2, 2, 8, 3),
 units = "micromoles")

#Implicit definition:
phr_reaction(Pyrite = 1, Reaction_amount = 5,
units = "millimoles", Linear_steps = 14)

# PHREEQC accepts not only EQUILIBRIUM_PHASES but any formula composed of
# defined SOLUTION_MASTER_SPECIES. It is up to you, however, to make sure
# the reaction you specify does actually make sense. In a REACTION block
# PHREEQC will dissolve anything, regardless of phase solubility

phr_reaction(C8H18O = 1, Reaction_amount = 0.1, Linear_steps = 10)


}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-33.htm#50528257_75635}
}
