% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input.R
\name{phr_reaction_temperature}
\alias{phr_reaction_temperature}
\title{Specify a reaction temperature gradient}
\usage{
phr_reaction_temperature(.number = 1, low = 0, high = 100, steps = 100)
}
\arguments{
\item{.number}{The number of the gradient}

\item{low}{The low temperature (degrees C)}

\item{high}{The high temperature (degrees C)}

\item{steps}{The number of steps to use between low and high}
}
\value{
A character vector summarising the input
}
\description{
Specify a reaction temperature gradient
}
\examples{
phr_reaction_temperature(low = 25, high = 75, steps = 51)

# example 2 from the manual
phr_run(phr_solution(), # pure water
        phr_equilibrium_phases(Gypsum = c(0, 1), Anhydrite = c(0, 1)),
        phr_reaction_temperature(low = 25, high = 50, steps = 10),
        phr_selected_output(temperature = TRUE, si = c("Gypsum", "Anhydrite")))

}
