% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_read_pqi}
\alias{phr_read_pqi}
\title{Read .pqi-files into tidyphreeqc}
\usage{
phr_read_pqi(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
An object of class \code{\link{phr_input}}
}
\description{
load .pqi-files produced by the PHREEQC-interactive GUI into
the tidyphreeqc format \code{\link{phr_input}}
}
\examples{
# create some phr_input_sections
sol <- phr_solution(pH = 8, pe = 2, Na = 1, Cl = 1, units = "mol/l")
phase <- phr_equilibrium_phases(Halite = c(10, 1))

fil <- tempfile("data")
phr_write_pqi(phr_input(sol, phase), path = fil)
phr_read_pqi(path = fil)
unlink(fil) # tidy up

}
