% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_run.R
\name{phr_run}
\alias{phr_run}
\alias{as.data.frame.phr_run_output}
\alias{as_tibble.phr_run_output}
\alias{print.phr_run_output}
\title{Run PHREEQC}
\usage{
phr_run(..., db = NA)

\method{as.data.frame}{phr_run_output}(x, ...)

\method{as_tibble}{phr_run_output}(x, ...)

\method{print}{phr_run_output}(x, ...)
}
\arguments{
\item{...}{Input components}

\item{db}{A database name, NA to use the default, NULL to use current phreeqc package db.}

\item{x}{Output from phr_run}
}
\value{
A list with components
}
\description{
Run PHREEQC
}
\examples{
phr_run(phreeqc::ex2)

}
