% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_selected_output}
\alias{phr_selected_output}
\title{Specify selected output}
\usage{
phr_selected_output(.number = 1, ...)
}
\arguments{
\item{.number}{Number of the component}

\item{...}{Further arguments}
}
\value{
A \link{phr_input_section}
}
\description{
Specify selected output
}
\examples{
phr_selected_output(pH = TRUE, pe = TRUE, alkalinity = TRUE, activities = c("H+", "OH-"))

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-38.htm#50528257_20239}
}
