% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_solution}
\alias{phr_solution}
\title{Sepcify a solution}
\usage{
phr_solution(.number = 1, .name = "", ...)
}
\arguments{
\item{.number}{Number of the component}

\item{.name}{Name of the component}

\item{...}{Further arguments}
}
\value{
A \link{phr_input_section}
}
\description{
Sepcify a solution
}
\examples{
phr_solution() # pure water
phr_solution(pH = 5.5, pe = 0)

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-41.htm#50528257_30253}
}
