% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_solution_list}
\alias{phr_solution_list}
\title{Create a large number of solutions}
\usage{
phr_solution_list(pH = 7, pe = 4, temp = 25, ...)
}
\arguments{
\item{pH}{the pH of the solution}

\item{pe}{the pe of the solution}

\item{temp}{the temperature of the solution (degrees C)}

\item{...}{Arguments to \code{\link[=phr_solution]{phr_solution()}}, all unique combinations of which will be used.}
}
\value{
A \code{\link[=phr_input]{phr_input()}}
}
\description{
Create a large number of solutions
}
\examples{
phr_solution_list(pH = c(5:7), pe = c(0:10), Hg = 0.1)

}
