% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_solution_species}
\alias{phr_solution_species}
\alias{phr_surface_species}
\title{Define a solution/surface species}
\usage{
phr_solution_species(reaction, log_k = NULL, delta_h = NULL,
  analytical_expression = NULL, ...)

phr_surface_species(reaction, log_k = NULL, delta_h = NULL,
  analytical_expression = NULL, ...)
}
\arguments{
\item{reaction}{The formation reaction of the species, e.g.: \code{CO3-2 + H+ = HCO3-}.}

\item{log_k}{The equilibrium coefficient of the reaction at 25 degrees C.
Alternatively can be specified for all temperatures using
\code{analytical_expression}.}

\item{delta_h}{The enthalpy of reaction, in kJ/mol (or with defined units).
Used to determine the temperature dependence of K according to the
Van't-Hoff equation, if \code{analytical_expression} is not provided.}

\item{analytical_expression}{Identifier for coefficients for an analytical
expression for the temperature dependence of log K. Must be a vector of six
numeric values: \code{A1 + A2*T + A3/T + A4*log10(T) + A5 / T^2 + A6*T^2}.}

\item{...}{Further arguments in the input block

A \code{\link[=phr_input_section]{phr_input_section()}}}
}
\description{
Define a solution/surface species
}
\examples{
# default units are kJ/mol for log_k
sp1 <- phr_solution_species("CO3-2 + H+ = HCO3-", log_k = 10.329, delta_h = -14.899)
# can also specify units
sp2 <- phr_solution_species("CO3-2 + H+ = HCO3-", log_k = 10.329, delta_h = "-3.561 kcal")
# log_k can also be specified through an analytical expression
sp3 <- phr_solution_species(
 "CO3-2 + H+ = HCO3-",
 analytical_expression = c(107.8871, 0.03252849, -5151.79, -38.92561, 563713.9, 0)
)

list(kJ = sp1, kcal = sp2, analytic = sp3, default = NULL) \%>\%
  lapply(
    phr_run,
    phr_solution("C" = "1 as CO3-2", temp = 90, units = "mol/L"),
    phr_selected_output(activities = "HCO3-")
  )

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-43.htm}
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/html/final-62.html}
}
