% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_surface_master_species}
\alias{phr_surface_master_species}
\title{Define a surface master species}
\usage{
phr_surface_master_species(binding_site_name, master_species)
}
\arguments{
\item{binding_site_name}{A name for the binding site, starting with a capital letter and including
only lowercase letters afterward (e.g., Mysurface). An underscore will denote an additional binding site on the
surface (e.g., Mysurface_one).}

\item{master_species}{The formula for the binding site. This is generall the binding site name
followed by OH (e.g., MysurfaceOH), in what the documentation calls the "OH-form of the binding site).

A \code{\link[=phr_input_section]{phr_input_section()}}.}
}
\description{
Defines a surface and a binding site on that surface (separated by an underscore). An identity
reaction with a log_k of 0 must be included in a \code{\link[=phr_surface_species]{phr_surface_species()}} block.
}
\examples{
phr_input(
  phr_surface_master_species("Mysurface", "MysurfaceOH"),
  phr_surface_species("MysurfaceOH = MysurfaceOH")
)

}
\seealso{
\url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/html/final-61.html}
}
