% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_tidy_PHREEQC}
\alias{phr_tidy_PHREEQC}
\title{Deal with input from a .pqi-file}
\usage{
phr_tidy_PHREEQC(x)
}
\arguments{
\item{x}{list of strings that constitute a .pqi-file}
}
\value{
A \code{\link{phr_input}}
}
\description{
Deal with input from a .pqi-file
}
\details{
This function is the working horse behind \code{\link{phr_read_pqi}}.
It takes in the list of character vectors read from the .pqi file,
determines when an where that list is to be split along the different
phreeqc keywords and then shoves the pieces into
\code{\link{phr_parse_PHREEQC}} to force meaning upon them.
It is not meant to be exported or to be used anywhere else, yet.
}
