% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_db.R
\name{phr_use_db}
\alias{phr_use_db}
\alias{phr_get_current_db}
\alias{phr_use_db_amm}
\alias{phr_use_db_ex15}
\alias{phr_use_db_iso}
\alias{phr_use_db_llnl}
\alias{phr_use_db_minteq}
\alias{phr_use_db_minteq.v4}
\alias{phr_use_db_pitzer}
\alias{phr_use_db_sit}
\alias{phr_use_db_wateq4f}
\alias{phr_use_db_phreeqc}
\title{Use various phreeqc databases}
\usage{
phr_use_db(db_string, save = TRUE, name = NA)

phr_get_current_db()

phr_use_db_amm(save = TRUE)

phr_use_db_ex15(save = TRUE)

phr_use_db_iso(save = TRUE)

phr_use_db_llnl(save = TRUE)

phr_use_db_minteq(save = TRUE)

phr_use_db_minteq.v4(save = TRUE)

phr_use_db_pitzer(save = TRUE)

phr_use_db_sit(save = TRUE)

phr_use_db_wateq4f(save = TRUE)

phr_use_db_phreeqc(save = TRUE)
}
\arguments{
\item{db_string}{A character vector version of the database.}

\item{save}{Pass TRUE to save the database as the default}

\item{name}{The (optional) name of the database being used.}
}
\value{
db_string, invisibly
}
\description{
For details on the actual databases, see \link[phreeqc]{Amm.dat},
\link[phreeqc]{ex15.dat}, \link[phreeqc]{iso.dat},
\link[phreeqc]{llnl.dat}, \link[phreeqc]{minteq.dat},
\link[phreeqc]{minteq.v4.dat}, \link[phreeqc]{pitzer.dat},
and \link[phreeqc]{sit.dat}. The default (loaded on namespace load)
is phreeqc.
}
\examples{
phr_use_db_phreeqc()
# equivalent to
phr_use_db(phreeqc::phreeqc.dat)

phr_use_db_amm()
phr_use_db_ex15()
phr_use_db_minteq()
phr_use_db_minteq.v4()
phr_use_db_pitzer()
phr_use_db_sit()
phr_use_db_wateq4f()

}
