% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_write_pqi}
\alias{phr_write_pqi}
\title{Write phr_input_sections to file}
\usage{
phr_write_pqi(..., .file, .addEND = FALSE, .overwrite = TRUE)
}
\arguments{
\item{...}{One or more phr_input_sections}

\item{.file}{Path to file}

\item{.addEND}{Logical. Shall a PHREEQC "END" command be appended add the end of the function call? This can be usefull if you are planning to write several models into a single file.}

\item{.overwrite}{Logical. If .file already exists, shall it be overwritten (the default) or sould new input be appended to the file?}
}
\description{
write your phr_input sections generated in tidyphreeqc to
.pqi files readily interpretable by the widely used PHREEQC-Interactive GUI
provided by the USGS to share your models with colleagues.
}
\examples{
# create some phr_input_sections
sol <- phr_solution(pH = 8, pe = 2, Na = 1, Cl = 1, units = "mol/l")
phase <- phr_equilibrium_phases(Halite = c(10, 1))

fil <- tempfile("data.pqi")
phr_write_pqi(sol, phase, .file = fil, .addEND = TRUE)
if(interactive()) file.show(paste0(fil,".pqi"))
unlink(fil) # tidy up

}
