% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_db.R
\name{use_db}
\alias{use_db}
\alias{use_db_amm}
\alias{use_db_ex15}
\alias{use_db_iso}
\alias{use_db_llnl}
\alias{use_db_minteq}
\alias{use_db_minteq.v4}
\alias{use_db_pitzer}
\alias{use_db_sit}
\alias{use_db_wateq4f}
\alias{use_db_phreeqc}
\title{Use various phreeqc databases}
\usage{
use_db(db_string, save = TRUE)

use_db_amm(save = TRUE)

use_db_ex15(save = TRUE)

use_db_iso(save = TRUE)

use_db_llnl(save = TRUE)

use_db_minteq(save = TRUE)

use_db_minteq.v4(save = TRUE)

use_db_pitzer(save = TRUE)

use_db_sit(save = TRUE)

use_db_wateq4f(save = TRUE)

use_db_phreeqc(save = TRUE)
}
\arguments{
\item{db_string}{A character vector version of the database.}

\item{save}{Pass TRUE to save the database as the default}
}
\value{
NULL, invisibly
}
\description{
For details on the actual databases, see \link[phreeqc]{Amm.dat},
\link[phreeqc]{ex15.dat}, \link[phreeqc]{iso.dat},
\link[phreeqc]{llnl.dat}, \link[phreeqc]{minteq.dat},
\link[phreeqc]{minteq.v4.dat}, \link[phreeqc]{pitzer.dat},
and \link[phreeqc]{sit.dat}. The default (loaded on namespace load)
is phreeqc.
}
\examples{
use_db_phreeqc()
# equivalent to
use_db(phreeqc::phreeqc.dat)

use_db_amm()
use_db_ex15()
use_db_minteq()
use_db_minteq.v4()
use_db_pitzer()
use_db_sit()
use_db_wateq4f()

}
