% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input.R
\name{phr_input_section}
\alias{phr_input_section}
\alias{new_phr_input_section}
\alias{as.character.phr_input_section}
\alias{as.character.phr_input_section_character}
\alias{print.phr_input_section}
\alias{is_phr_input_section}
\alias{as_phr_input_section}
\alias{as_phr_input_section.phr_input_section}
\alias{as_phr_input_section.character}
\title{Create input sections for phreeqc}
\usage{
phr_input_section(type, number = NA, name = "", components = list())

new_phr_input_section(x)

\method{as.character}{phr_input_section}(x, ...)

\method{as.character}{phr_input_section_character}(x, ...)

\method{print}{phr_input_section}(x, ...)

is_phr_input_section(x)

as_phr_input_section(x, ...)

\method{as_phr_input_section}{phr_input_section}(x, ...)

\method{as_phr_input_section}{character}(x, ...)
}
\arguments{
\item{type}{The keyword name to use (e.g. SOLUTION)}

\item{number}{Number of the component}

\item{name}{Name of the component}

\item{components, ...}{Key/value pairs corresponding to lines of input}

\item{x}{An object created by phr_input_section}
}
\value{
A character vector with an element for each line
}
\description{
Create input sections for phreeqc
}
\examples{
# pure water solution
phr_solution(pH=7, temp=25)

}
