% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input_helpers.R
\name{phr_reaction_temperature}
\alias{phr_reaction_temperature}
\alias{phr_reaction_pressure}
\title{Specify a reaction temperature, pressure gradient}
\usage{
phr_reaction_temperature(.number = 1, low = 0, high = 100, steps = 10,
  values = NULL)

phr_reaction_pressure(.number = 1, low = 0.5, high = 1.5, steps = 10,
  values = NULL)
}
\arguments{
\item{.number}{The number of the gradient}

\item{low}{The low temperature, pressure (degrees C, atm)}

\item{high}{The high temperature, pressure (degrees C, atm)}

\item{steps}{The number of steps to use between low and high}

\item{values}{Explicit values to use (takes precedence over low, high, steps)}
}
\value{
A character vector summarising the input
}
\description{
Specify a reaction temperature, pressure gradient
}
\examples{
phr_reaction_temperature(low = 25, high = 75, steps = 51)
phr_reaction_temperature(values = c(12, 15, 25))

# example 2 from the manual
phr_run(
  phr_solution(), # pure water
  phr_equilibrium_phases(Gypsum = c(0, 1), Anhydrite = c(0, 1)),
  phr_reaction_temperature(low = 25, high = 50, steps = 10),
  phr_selected_output(temperature = TRUE, si = c("Gypsum", "Anhydrite"))
)

}
\seealso{
Temp: \url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-35.htm#50528257_75016}
Pressure: \url{https://wwwbrr.cr.usgs.gov/projects/GWC_coupled/phreeqc/phreeqc3-html/phreeqc3-34.htm#50528257_65966}
}
