% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_run.R
\name{phr_run}
\alias{phr_run}
\title{Run PHREEQC}
\usage{
phr_run(..., db = NA, dump = NA)
}
\arguments{
\item{...}{Input components}

\item{db}{A database name, NA to use the default, NULL to use current phreeqc package db.}

\item{dump}{Optional flag to include dump information. Defaults to NA, which detects whether
a DUMP block is included in the input.}
}
\value{
A list with components
}
\description{
Run PHREEQC
}
\examples{
phr_run(phreeqc::ex2)

}
