% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_input.R
\name{phr_solution_list}
\alias{phr_solution_list}
\title{Create a large number of solutions}
\usage{
phr_solution_list(pH = 7, pe = 0, ...)
}
\arguments{
\item{pH}{the pH of the solution}

\item{pe}{the pe of the solution}

\item{...}{Arguments to \link{phr_solution}, all unique combinations of which will be used.}
}
\value{
A \link{phr_input}
}
\description{
Create a large number of solutions
}
\examples{
phr_solution_list(pH = c(5:7), pe = c(0:10), Hg = 0.1)

}
