% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phreeqc_write_pqi.R
\name{phr_write_pqi}
\alias{phr_write_pqi}
\title{Write phr_input_sections to file}
\usage{
phr_write_pqi(x, path)
}
\arguments{
\item{x}{An object of class "phr_input". See \code{\link[=phr_input]{phr_input()}} for details}

\item{path}{Path to file.}
}
\value{
\code{x}, invisibly
}
\description{
Write a \code{\link[=phr_input]{phr_input()}} generated in tidyphreeqc to
.pqi files readily interpretable by the widely used PHREEQC-Interactive GUI
provided by the USGS to share your models with colleagues.
}
\details{
The path supplied to the function might typically look something
like this: "~/path/to/my_program.pqi" but note that the ".pqi" ending is
strictly optional and any (or no) file-ending is accepted. However, in order
to be automatically recognised by the PHREEQC-Interactive GUI as a
PHREEQC-input file, the ".pqi" ending is necessary.
}
\examples{
# create some phr_input_sections
sol <- phr_solution(pH = 8, pe = 2, Na = 1, Cl = 1, units = "mol/l")
phase <- phr_equilibrium_phases(Halite = c(10, 1))

fil <- tempfile("data")
phr_write_pqi(phr_input(sol, phase), path = fil)
unlink(fil) # tidy up
}
